requirements.txt
pandas
chromadb
sentence-transformers
groq
python-dotenv
tqdm